pyproject.toml
[tool.poetry]
name = "drf-ecommerce-project"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.7"
django-filter = "^23.4"
djangorestframework = "^3.14.0"
drf-spectacular = "^0.26.5"
pytest = "^7.4.3"
pytest-django = "^4.7.0"
python-decouple = "^3.8"
pytest-factoryboy = "^2.6.0"
django-mptt = "^0.15.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
